requirements.txt
opencv-python>=4.8.0
mediapipe>=0.10.0
numpy>=1.24.0
pyautogui>=0.9.54
Pillow>=10.0.0
screeninfo>=0.8.1

